% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WassPWdense.R
\name{WassPWdense}
\alias{WassPWdense}
\title{Wasserstein pairwise warping for completely/densely observed distribution-valued random processes}
\usage{
WassPWdense(tVec, qf, qSup, optns = list())
}
\arguments{
\item{tVec}{A vector holding the common equidistant time grid on which all \eqn{\{X_i(\cdot)\}_{i=1}^{n}} are observed.}

\item{qf}{A list of matrices with \code{qf[[i]]} holding the \eqn{i}-th distributional 
trajectory \eqn{X_i(\cdot)} on \code{tVec} represented by quantile functions on \code{qSup}, 
of which each row corresponds to a time point in \code{tVec}.}

\item{qSup}{An increasing grid between 0 and 1 holding the support grid on which the quantile functions are evaluated.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}, 
including all those available in \code{\link[fdapace]{WFDA}} except for that the default \code{subsetProp} is 1.}
}
\value{
A list of the following:
\item{h}{A matrix holding the warping functions evaluated on \code{workGrid}, each row corresponding to a subject.}
\item{hInv}{A matrix holding the inverse warping functions evaluated on \code{workGrid}, each row corresponding to a subject.}
\item{qfAligned}{A list of matrices. The \eqn{i}-th matrix holds the \eqn{i}-th aligned 
distributional trajectory on \code{workGrid} represented by quantile functions on \code{qSup}, 
where each row corresponds to one time point in \code{workGrid}.}
\item{qSup}{An increasing grid between 0 and 1 holding the support grid on which the quantile functions are evaluated.}
\item{workGrid}{A copy of \code{tVec}. An equidistant grid on which 
(inverse) warping functions in \code{h} and \code{hInv} 
and distributional trajectories in \code{qf} and \code{qfAligned} are evaluated.}
\item{optns}{Control options used.}
\item{costs}{The mean cost associated with each trajectory.}
\item{timingWarp}{The time cost of time warping.}
}
\description{
Pairwise warping for distribution-valued random processes 
\eqn{\{X_i(\cdot)\}_{i=1}^{n}} endowed with 2-Wasserstein metric, 
where each \eqn{X_i(\cdot)} is a distribution-valued random process densely observed
 on the same grid across \eqn{i} (and hence no local Fréchet regression is involved).
}
\references{
\cite{Chen, Y., and Müller, H.-G. (2023). "Uniform convergence of local Fréchet regression with applications to locating extrema and time warping for metric space valued trajectories." The Annals of Statistics, 50(3), 1573--1592.}

\cite{Tang, R. and Müller, H.-G. (2008). "Pairwise curve synchronization for functional data." Biometrika 95, 875--889.}
}
